Introduce a (pomless) tycho build for epic
This makes building epic a lot easier than the ant-based build scripts,
mostly because it requires 'just' maven to be present on the systems to
build the project. Support for having maven available is often integrated
in CI solutions already - unlike a headless Eclipse install that is needed
to drive the ant-based build.

There is one drawback: The site.xml file had to be renamed to category.xml
as that's what Tycho is expecting as the name for the upate site
configuration. I don't know if this impacts the ant-build but it's likely
it does.

The build is done against Eclipse 4.8, with Java 1.6 as target (since that's
what the configuration files still refer to) and ECF 3.14.1 (because it's
released around the time of Eclipse 4.8). Tycho/Pomless is set to 1.5.1
which seems to be the last version supporting 'arbitrary' directory layouts
for the plugins. Newer versions complain about not being able to identify
the folders, likely because they require the 'structured build' structure
used by Eclipse itself as well.

Added 'target/' to a gitignore file so that the output of maven is ignored
by git.