Add a simple script to generate the perl parser from the antlr grammar
The eclipse code requires this step, for now I'm planning to do this outside
of maven as it is easier implemented that way. Unfortunately a rather
old version of antlr is being used, which does not have a maven plugin.
Luckily the necessary binary jar is provided as part of the project so it's
easy to script this.