Convert run.gif to run.png in the documentation
This plugin is extracted when installed into a product, leading to a
'run.gif' file to be present in the product folder. Various virus scanners
consider any 'run.gif' files a potential threat and block unpacking of
a product's zip file when it has such a file inside.

Using png as file type has no such problem, so change the type for the docs.

I left the gif's in the debug and regexp plugin as is as those are
distributed inside the jar and that does not seem to be problematic.